Add support for trinket tinkering, Focusing Lenses and Horns of Leadership. Update vendor NPCs (#348)
* Add support for trinket tinkering and horn of leadership. Update vendor NPCs

This is a working PR. I will add support for the focusing lenses, medals and virindi essence devices at a later date.

* Fix outdated Killtask on 44019 War Armoredillo

Removed out dated kill task emote on 44019 and removed uneeded blank space from a few other dillos.

* Update Recipes for IconUnderlay Imbued Effect

Added the recipe mod to add the Icon underlay to the imbued trinkets themselves. Changed icon underlay for pyreal and amber to align with retail.

* Add Spells and Adjust skill check on horns

Moved the horned to correct directory and adjusted thier skill checks. Added spells to the spell table.

* Added Focusing Lenses and Recipes, Updated Trinket Imbue Spells, Updated Jewelers Inventories

All the focusing lense stuff is in now. Had bad data in trinket imbue spells, had to fix that and updated all/most jewelers to support selling the lense parts for tinkering.

* Update 8478 Flawless Lense.sql

orrect recipe to carved pyreal handle

* Update ItemCurMana to match ItemMaxMana on finished lenses

* Set lenses ItemManaCost

Set INT 117 to 100 for ItemManaCost on lenses.