Update Neftet and FV Turn In NPC Emotes
Update emotes so when turning in the quest items before speaking with the NPCs players will not revieve an infinite quest timer response. Instead the NPC's will ask the player to start a conversation first.

This will cut down on bug reports until more time is invested in re-writing the emotes further.