Add Amelia's Lost Toys quest
The basic flow of the quest is that the player talks to Amelia (35932) which
pops up a yes/no box. If the player accepts then a signal is sent to spawn
Amelia (35933), then Amelia (35932) despawns. Amelia will then run to
certain points and emit a signal.

There are generators setup at each location which listen for
the signal and then spawn a specific dirt pile. The dirt pile last for 30
seconds and then disappear. When the player interacts with the dirt pile
it will give them a specific toy.

After going to the last dirt pile Amelia (35933) will run home, emit a
signal to spawn the original version (35932), then despawn.

Amelia (35932) will accept the toys and give a reward once all 6 toys have
been handed over.

Information on dirt pile locations was taken from PCAP but also from this
video: https://www.youtube.com/watch?v=9SyAqFNuwzo. Not all locations were
shown in the video so some guesswork was done.