update to PreTOD Guises (#859)
Ive updated the clothing priority and valid locations. to reflect end of retail. Wiki shows screen shots of Full Coverage in these guises. where as ACE core weenies did not cover hands. I believe these were updated sometime in 2006 but I could not find any mention of it on the Wiki, or Fandom.
http://acpedia.org/wiki/Full_Mu-miyah_Guise
http://acpedia.org/wiki/Scarecrow_Guise
http://acpedia.org/wiki/Armored_Undead_Guise
http://acpedia.org/wiki/Armored_Skeleton_Guise

Gmriggs suggested it go in the EOR folder.