update loot, jaw and shroud drop rates
Jaw drop rate now 1, 0.75 and 0.4 for pyre champions, skeletons, and minions, respectively.
Pyre shroud drop rate now 1 (with a 0.3 chance of a 2nd shroud) for pyre champions, and 0.5 for skeletons and minions.
Corrected XP on Pyre Champions so it is uniform.
Updated loot profiles on skeletons so only night spawns drop T8.