Uber Lady Aerfalle Quest
Flat version of the PR, with corrected crushing blow property on Reforged Ashbane.

The update to Kuziya bint Zayi the Translator is now in a separate PR 1164 and is required for Aerfalle's token rewards.

Added the uber version of Lady Aerfalle.
Added XP tokens to normal and hard version chests.
Updated normal and hard version global to refer to Aerfalle's Apprentice instead of Aerfalle.
Updated 06771 Leikotha rewards for Ashbane to end of retail, and added Reforged Ashbane.
Added int 353 WeaponType Magic to old versions of Staff of Aerfalle.
Updated Sacrificial Dagger to end of retail stats.
Added 40932 Ghost of Dylaeral to handle easy version since 28060 Ghost of Galaeral now handles the uber version.

Note: The enhanced versions of Atlan and Isparian Weapons are NOT included in this PR.