Update Arcane Death
Increased the damage on the Arcane Death spell used for traps in the Jester, Empyrean Rescue and Hoshino Must Die.

Sample damage numbers on a player with pierce pro 8, legendary pierce ward, and old school aegis equipped:

412, 455, 425, 486, 495, 474, 454, 439, 403, 427

Updated name of Hoshino Must Die traps to Hall of Agony as seen in a retail video.

Updated comments in the dungeon landblock with the traps to match.