revised gens
Based on another retail video, now using a single gen for gromnie and gromnus, since a gromnus could spawn in the same location as a gromnie.
Slightly moved some gens that could not generate a gromnus due to not enough space.
Corrected slithis names so no double apostrophy, and added missing spell to slithis tentacle.