Update 43808 Carefully rolled scroll.sql (#1318)
removed pickup timer. There was no pickup timer for this scroll. Ran the quest alot. When I picked up the first scroll, go turn in, come back thru and you could pick up another with no timer. Besides why have a timer on the scroll, it does nothing but move you to the next step of the quest, no way to exploit or abuse.