Rynthid Quest Fixes
Rynthid Assessment Crystal

Fixed function and appearance of training dummies based on pcap and retail screenshots.
Added missing magma to one of the jump test rooms as seen in retail screenshot.
Corrected message when player steps in a corrupted energy as seen in retail screenshot.
Added no recall property to 2 colored portals that were missing it, to match the other colored portals and retail (player should not have been able to portal recall to return to a trial if they died).

Crimson Scarab

Corrected appearance of Sparking Crystal so it matches retail video.
The sparking crystals now also spawn 2 lightning elementals when approached, visible on retail teaser image. This is also consistent with the video - in the video, player does not approach crystals, therefore these lightning elementals are never seen.