Armor of Darkness Quest
Adds the Armor of Darkness quest:

http://acpedia.org/wiki/Armor_of_Darkness_Quest

Note:

This PR updates the Hebian To landblock E74E.

This PR does not include the Gear Knight Parts quest, which will be added separately.

Known issues:

Monster balance and appearance are a work in progress.
The Unknown Warrior does not attack when you talk to him. Must attack him first.
Some of the potion icons may be incorrect.