Update Lunnum's Return
Added missing doors and portal from vlocs to the blue face paint recipe dungeon. Also added some more mobs and crates based on the single retail screenshot.

http://acpedia.org/wiki/Storage_Crate_(Nuhmudira%27s_Dungeon)

Changed blue face recipe to a generic misc item and corrected it so all info would display in the assess panel. Previously, it was a writable, but completely blank.

Increased spawn in the temple dungeon to 5 min. It had been left with the default generator from the pcap, and the retail respawn was not that slow.