revisions
LadyJaeraWandPickup replaced by TrophyMageAcademy_1110 to match contract.
Added contract for the quest.
Edited emote flavor text for Rytheran's Apprentice so it matches the fact that this is an apprentice, not Rytheran himself.
Added new Sheets of Paper to the apprentice.
Updated Zaikhal translator to take the Sheets of Paper and give translation.
Corrected Lord Kelannik's art.
Corrected Tazal al-Ashfai's level, and typo in his emote.