Flavor Text and Other Updates
Adds on death local broadcast to Forge Golem seen in retail video by Xeo, and updates the golem's skills and casting probability. Some of its base skills were set to what should be its effective skills.
Adds flavor text to Forges portal seen in that retail video.
Updates use range on Forges and Armory portals to fix issue of player walking around in circles until finally being able to get in the portal. In the retail video, Xeo was simply able to walk to the center of the portal platform and get portaled.
Adds a regular door reset interval of 300 seconds to regular (not lever activated) Neftet-style door. This was not set, and doors closed too quickly  (in under a min). In retail video, these doors stay open for much longer.
Updated Rynthid Testing portal generator to match the other portal generators. The others are set to 5 sec regen interval, this one was left at 600 seconds, forcing players to wait around for 10 min for a new color portal to spawn.
Added missing T7 loot profile to Verdant Moarsman.