Bulb of Mornings Updates
Revises the Harvesting the Bulb of Mornings quest based on Loud Lou retail video:

Updated Kaltus emote text to match retail from the video (parts are missing from the wiki).
Some mob positions adjusted, and missing mobs added.
Corrected art for some of the reedsharks.
Adjusted some mob stats and resistances. Some of the stats are incorrect on the wiki, but seen on the video.
Removed phantom Bulb of Mornings guid from the landblock (this appeared to be causing a duplicate bulb to spawn), and adjusted bulb emote.
Flowers are now static spawns as seen on the video (i.e., always respawn in the same places on a linkable).
Corrected exit portal drop.