Climbing the Spire Updates
Revisions to the Shield of Isin Dule quest based on pcap and ImmortalBob retail video:

Corrects Shadow Cyst spell book based on what it casts in the video. The spells are just level 7 versions of the same spells the low level Shadow Cyst (from world db) casts, with a few replacements.

Adds missing mobs the Shadow Cyst can generate. In the pcap, there is Isin Dule's Lieutenant, and in a retail screenshot, there is Isin Dule's Captain.  Neither appear in the video. This could be because dealing enough damage might skip over the health ranges to generate them (if this was based on wounded taunt).

Adjusts Shadow Cyst resistances (using those for the world Shadow Cyst, with higher armor level). ImmortalBob deals very high damage (over 6k on criticals) prior to the Shadow Cyst casting fire protection on itself (reducing damage to over 2k on criticals).

Corrected scale on Shadow Cyst to match the pcap (should be a scale of 1).

The wielded treasure profile (5920) used for some of the soldier type shadows appears to be broken, causing them to spawn unarmed. I used the create list to accomplish the same thing (1 of 5 possible swords) and shield equipped.

Reduced wave generator delay to 1200 (was 1600 previously), to speed up respawn of the shadows needed to enter the spire. It is about 20 min between when ImmortalBob kills the first wave to when he exits back out at the top of the spire. The shadows below can be seen to have already respawned by then.