Essence Chambers Updates
Revises Essence Chambers based on a Loud Lou retail video:

Cleaned up Hands of Enchantment that should only spawn when Essence of Enchantment reaches 50% HP.
Added generate Hands for all Essences at around 50% HP.
Revised to remove events, replacing them with a wave generator.
Added Lair to the Harbinger portals (4 of them) that spawn when all of the Essences die and last for 5 min.
Removed transparency from the Hands to correct their appearance, matching the pcap (and the video).
Revised resistances on the essences for a closer match to the video, and set armor value from the pcap.