Rynthid Testing Fix Broken Reset (#1539)
Currently, the door tests in this quest only reset when the landblock resets, meaning players are stuck waiting for a respawn that will never come. This corrects this issue by having the door controller stopgap delete itself after it closes its door, allowing the generator it is on to start its cycle from the beginning.

Drudge Wilter scale is adjusted to match retail video.

Gurog Bruiser stats and art are adjusted to match retail video. These Gurog do not appear to ever cast and only melee with their fists. They do not wield an axe. Also, their HP was lower than it should be based on how much damage it takes to kill them on the video.

No corpse is set on both of these mobs, and loot profile removed, as seen on the video.