Bur Fixes (#1563)
Revises skill levels on several boss creatures on Bur, mainly part of the Saving Asheron quest, to match pcap estimated levels.

Mana Barrier and Blooded Ritual Statues are corrected based on a Berek retail video. Previously, they had been set to resist a max player about 50% of the time. However, there is no evidence of this in the video. Berek lands spells 100% of the time, in the case of the mana barrier, even after his buffs are dispelled by a trap.  Melee and missile defense on the statue are also set to negligible levels to match are immobile objects (that could be machine gunned). The melee defense pcap skill estimate is 2 (i.e., almost 0) for these.

Fixed the mana barrier trying to attack the player, and spinning around like revolving door, making it useless as a barrier.

As the video shows, the barrier and statues had a lot more health than 10k. It is now set to 50k.

Set creature type for some Gate Guardians that were missing it (should be undead).

Corrected properties on Corpse of Servitor of Asheron (so it appears like object), and what appear to be player-made typos on the wiki (I could not find evidence that these were Turbine-made typos in the pcap).

Added missing hotspots to the leyline (as described on the wiki, and as I remember them from retail), they deal a lot of damage if the player stands too close.

Adds missing NPC appears like object to Kresovus' Lockbox.