V0.9.271 (#1564)
* updated to match previous dropped wcid (#1559)

* Add Olthoi Pauldrons to VR recipes (#1560)

* Fix Part 5 of Shadow Invasion increment for Ler Rhan (#1561)

* Dark Isle Updates (#1562)

* Dark Isle Updates

Revises skills (based on pcap skill estimates) and resistances (based on available video) of Dark Isle creatures. In general, their magic skills were too high, while melee and missile skills were too low. Also adds missing shadow types (those that look like a viamontian and a royal knight).

* Update Golem Casting

* Update 33733 Depraved Mukkir.sql

* Update 40282 Depraved Mukkir.sql

---------

Co-authored-by: Ty Conner <tyconner@itcproductions.com>

* Bur Fixes (#1563)

Revises skill levels on several boss creatures on Bur, mainly part of the Saving Asheron quest, to match pcap estimated levels.

Mana Barrier and Blooded Ritual Statues are corrected based on a Berek retail video. Previously, they had been set to resist a max player about 50% of the time. However, there is no evidence of this in the video. Berek lands spells 100% of the time, in the case of the mana barrier, even after his buffs are dispelled by a trap.  Melee and missile defense on the statue are also set to negligible levels to match are immobile objects (that could be machine gunned). The melee defense pcap skill estimate is 2 (i.e., almost 0) for these.

Fixed the mana barrier trying to attack the player, and spinning around like revolving door, making it useless as a barrier.

As the video shows, the barrier and statues had a lot more health than 10k. It is now set to 50k.

Set creature type for some Gate Guardians that were missing it (should be undead).

Corrected properties on Corpse of Servitor of Asheron (so it appears like object), and what appear to be player-made typos on the wiki (I could not find evidence that these were Turbine-made typos in the pcap).

Added missing hotspots to the leyline (as described on the wiki, and as I remember them from retail), they deal a lot of damage if the player stands too close.

Adds missing NPC appears like object to Kresovus' Lockbox.

* normalize

* Update 45493 Rare Coin.sql

* Update 45436 Bearded Axe of Souia-Vey.sql

* Update 45437 Canfield Cleaver.sql

* Update 45438 Count Renari's Equalizer.sql

* Update 45439 Smite.sql

* Update 45440 Tusked Axe of Ayan Baqur.sql

* Update 45441 Black Thistle.sql

* Update 45442 Moriharu's Kitchen Knife.sql

* Update 45443 Pitfighter's Edge.sql

* Update 45444 Ridgeback Dagger.sql

* Update 45445 Zharalim Crookblade.sql

* Update 45446 Baton of Tirethas.sql

* Update 45447 Dripping Death.sql

* Update 45448 Star of Tukal.sql

* Update 45449 Subjugator.sql

* Update 45450 Thunderhead.sql

* Update 45451 Champion's Demise.sql

* Update 45452 Pillar of Fearlessness.sql

* Update 45453 Squire's Glaive.sql

* Update 45454 Star of Gharu'n.sql

* Update 45455 Tri-Blade Spear.sql

* Update 45456 Staff of All Aspects.sql

* Update 45457 Death's Grip Staff.sql

* Update 45458 Staff of Fettered Souls.sql

* Update 45459 Spirit Shifting Staff.sql

* Update 45460 Staff of Tendrils.sql

* Update 45461 Brador's Frozen Eye.sql

* Update 45462 Defiler of Milantos.sql

* Update 45463 Desert Wyrm.sql

* Update 45464 Guardian of Pwyll.sql

* Update 45465 Morrigan's Vanity.sql

* Update 45466 Fist of Three Principles.sql

* Update 45467 Hevelio's Half-Moon.sql

* Update 45468 Malachite Slasher.sql

* Update 45469 Skullpuncher.sql

* Update 45470 Steel Butterfly.sql

* Update 40147 Black Coral Golem Viceroy.sql

Fix gen infinite loop.

* Update 40147 Black Coral Golem Viceroy.sql

* Update 45493 Rare Coin.sql

* Fixup el'yst (#1565)

* Update 49584 El'Yst Maggann.sql

* Update 49584.es

* finalize

---------

Co-authored-by: Schneebs <27392240+Schneebs@users.noreply.github.com>
Co-authored-by: Boop-AC <90673082+Boop-AC@users.noreply.github.com>
Co-authored-by: TectonicRifts <60861066+TectonicRifts@users.noreply.github.com>
Co-authored-by: ziang4891 <56269610+ziang4891@users.noreply.github.com>