geraine host fix
http://acpedia.org/wiki/Geraine%27s_Hosts

Note: In order for this quest to count as one of the Legendary Key Quests you must turn in all 7 books to the Mhoire Lieutenant in Wai Jhou.

you were able to get credit for this quest by turning in the 1st 4 books into the soldier. which is not retail