V0.9.274 (#1583)
* fluff ints added (#1572)

added unenchantable and weild requirements. basically fluff stuff. keeps the id panel from being empty.

* add delete self (#1573)

in retail if you had a corrupter amber to the fountain it deleted itself

* Fix Atamarr emotes (#1574)

* Stormwood Staff Spell Fix (#1575)

* Fix spells on Seasoned Explorer Nether Staff (#1576)

* Fix Spectral Blade and Claws Contract (#1577)

* Fix Red Rune Silveran Mace TurnIn (#1578)

* Fix deru tree royal olthoi jelly (#1579)

* Deru Tree lockout fix (#1580)

* Fix Nimble Fingers buffs (#1581)

* fix spell on t5 rare Hieroglyph of Healing Mastery, from Incantation of Heal Self -> Incantation of Healing Mastery Self (#1582)

* normalize

* Fix Apostate Finale portal restrictions (#1584)

* Fix 575 LP Viridian Statues (#1585)

* geraine host fix (#1586)

http://acpedia.org/wiki/Geraine%27s_Hosts

Note: In order for this quest to count as one of the Legendary Key Quests you must turn in all 7 books to the Mhoire Lieutenant in Wai Jhou.

you were able to get credit for this quest by turning in the 1st 4 books into the soldier. which is not retail

* normalize

* More statue updates

* normalize

* finalize

---------

Co-authored-by: Fire-Chimp <49045665+Fire-Chimp@users.noreply.github.com>
Co-authored-by: Boop-AC <90673082+Boop-AC@users.noreply.github.com>
Co-authored-by: gmriggs <gmriggs@gmail.com>