Update Nexus (#1624)
* Update Nexus

Updates the Nexus landblock 0110 to match the pcap and retail video by Tzepesh. This corrects spawn placement, and adds the missing traps (present in the original landblock from world DB, and seen in the video, meaning they were not removed). Also adds damage rating to Nexus Commander (as in the video).

https://www.youtube.com/watch?v=0W13OWw8SkU

* Update 0110.sql

Increased level of lightning trap to 7.