Update 37585 Soul Bound Staff.sql
Adjusting critical multiplier and frequency on the SB staff for a closer match to retail.

Based on a retail video of Berek fighting Aerbax, with pierce vul 8, the average on normal hits was 1317 (based on 66 shots). The critical average was 4432 (based on 11 shots). This is with pierce vul 8.

Without pierce vul (only on the staff's pierce rend), the normal hit average is 1031 (total 40 shots) and 3506 on criticals (total 10 shots).

Either way, criticals are about 3.5 more damage than normal hits. Also, out of the 149 shots in the video, 25 were criticals, which is a .1677 rate. Subtracting 1% for eye of remorseless aug, and rounding down to be more conservative, gives 15% chance to critical.