Update Rynthid Outdoors
Adjusts skills on rynthid to match the pcap skill estimates. In general, magic was too high, and in some cases melee and missile were too low.
Revises outdoor generators for a better match to retail based on Loud Lou video. From the video, I estimated respawn to be longer than 5 min but less than 10, so set to 7 min.
Fixes Rynthid Crystal so it will generate adds.