Fix Prodigal Lugian (#1686)
Corrects a number of issues with the Prodigal Lugian quest reported by players. In retail, multiple people could do this quest without interfering with each other. As is, the quest only works properly for one person at a time, and can cause others to get stuck for over 30 min due to unnecessary use of events and landblock keep alives.

This redesigns the quest, allowing multiple people to work on the trials at the same time (such as in the Burun Kings quest where a similar mechanic is used).

Also resolves console errors - missing combat table on a boulder, and failing to spawn weenie, which had not been included.
Corrects energy cage mechanic, and scene at the end of one trial.
Corrects final fight. As seen in the pcap, adds are clearly spawned by Kresovus around him, thus removing multiple unnecessary wave gens.
Adds missing emote to boulder in Trial of the Arm, which in retail, could be "moved" if the person had a high enough base strength (thus, the name of the trial). This is why this boulder is useable in the pcap. If strength was not high enough, it could alternatively be killed.
Fixes a bug where the Trial of the Mind was purely cosmetic (i.e., did not stamp the player correctly). The player could previously move on without doing the trial at all.
Cleans up duplicate spawns in the fur room.
Adds NPC looks like object to Altar.