Update Gurog Creation (#1692)
Minor updates and bug fixes to this quest based on retail videos by Kala Glow and Geffers58.

The Journal of a Soldier can now be read.
The Mortified Soldier is now in a ShiverState as seen in the video.
Set location of two doors from vlocs (these were not papped, and were previously manually placed).
Adjusted some spawns in non-pcapped sections of the dungeon to match the video.
Respawn increased to 5 min (was previously 3 min). The video shows it was longer than 3 min.