Misc Fixes October
Fixes dual wielding ghost weapons based on Loud Lou retail video. Previously, it was possible for a ghost weapon to spawn unarmed or wield only in its left hand, which was incorrect.

Corrects undefined damage type on bow wielding ghost weapons. It is now set to pierce, as seen in the video.

Sets scale of Summoning Chamber Adept to 1 to match retail screenshot.