Update Hoshino Fort Main Gate
In retail, the main gate of Hoshino Fort would stay open for a longer period of time (around 30 min), as far as I remember, even when no one was around to keep the landblock active.

To replicate this behavior, when activated, the lock now spawns a Landblock KeepAlive with a lifespan of 25 min, and the reset interval on the gate is set to 30 min.