Update Neftet Generators (#1711)
Added Bloodstone to Neftet generators, and revised spawns to eliminate redundancies, and excessive amount of landscape trophies.

Deleted two unnecessary generators.

The trophies and special spawns (tablets, cacti, etc.) are now found only in the master generators. The exception is desert flowers, which can also spawn in the armoredillo camps, as seen in retail videos.