Update BDC Quiddity Seed (#1755)
* Update BDC Quiddity Seed

Updated the quest to fix issues, and added missing elements to match retail video by Tzepesh.

Most strange devices now function without events or local signals. Deleted the events that were replaced. The exception is the device at Asheron's Castle since the scene takes place across two landblocks.

Added missing Aluvian Rebels to the 2nd Underground City scene.

Previously, the player only had 30 seconds to pick up the head of the olthoi queen at the end, or it disappeared, which failed the quest forcing a 20 hour wait. The head now persists until it decays normally.

Filled in more of the BDC dungeon with mobs. Most of it had been left empty since only the quest path was pcapped.

Added missing stats to Elysa, Thornsten and the Olthoi Queen.

Corrected a typo in the Quiddity Seed dialog. This typo was not in the retail video text.

* elysa and thorsten stats