V0.9.287 (#1754)
* Fix some edge cases with emote prompts (#1752)

* Add missing weapon type (int 353) to Tusker Bone Sword (#1753)

* Update BDC Quiddity Seed (#1755)

* Update BDC Quiddity Seed

Updated the quest to fix issues, and added missing elements to match retail video by Tzepesh.

Most strange devices now function without events or local signals. Deleted the events that were replaced. The exception is the device at Asheron's Castle since the scene takes place across two landblocks.

Added missing Aluvian Rebels to the 2nd Underground City scene.

Previously, the player only had 30 seconds to pick up the head of the olthoi queen at the end, or it disappeared, which failed the quest forcing a 20 hour wait. The head now persists until it decays normally.

Filled in more of the BDC dungeon with mobs. Most of it had been left empty since only the quest path was pcapped.

Added missing stats to Elysa, Thornsten and the Olthoi Queen.

Corrected a typo in the Quiddity Seed dialog. This typo was not in the retail video text.

* elysa and thorsten stats

* Update Ithaenc Quiddity Seed (#1756)

Corrected an issue with the Shade of Dule emote where the on use text was missing if the player was in a fellow.

Added the standard "you cannot read the writing on this" text on the untranslated Thief of Dreams Message Shard.

* Update Gronk and Tanada Neophyte (#1757)

Corrects the appearance of Gronk the Hammer - thanks Raile for the fix.

Corrects scale of Tanada Neophyte. They were a little smaller compared to the other Tanada. Thanks Fire Chimp for the report.

* normalize

* Update README.md (#1758)

Spelling corrections.

* Update Ayaname Kina and Deewain (#1759)

* Update Ayaname Kina and Deewain

Adds missing girth armor to Ayaname Kina.
Updates Deewain dialogue to the correct one at his final location on Asheron's Island.

* Update Ayaname Kina

Adds interaction for 73246 Purified Quiddity Ingot from PR 1760.

* Add BDC Quiddity Ingot Purification (#1760)

Adds this quest:

https://asheron.fandom.com/wiki/Black_Death_Catacombs_Quiddity_Ingot_Purification

* Add Ithaenc Quiddity Ingot Purification Quest (#1761)

* Add Ithaenc Quiddity Ingot Purification

Adds the quest:

https://asheron.fandom.com/wiki/Ithaenc_Quiddity_Ingot_Purification

Previously, the dungeon was in but unfinished.

* Create IthaencQuidityIngotPurified.sql

* Update 73247 Well of Blood.sql

Adjust scale.

* Add Knorr Quiddity Ingot Purification (#1762)

Adds the quest:

https://asheron.fandom.com/wiki/Knorr_Quiddity_Ingot_Purification

* normalize

* finalize

---------

Co-authored-by: Boop-AC <90673082+Boop-AC@users.noreply.github.com>
Co-authored-by: TectonicRifts <60861066+TectonicRifts@users.noreply.github.com>
Co-authored-by: Galli <99272482+KT053374@users.noreply.github.com>