update guardian of the deru (#1766)
Revises dungeon spawns for a closer match to retail (sclavus in areas with olthoi corpses, olthoi in organic areas).

All tentacles of tthuun outside now count for the kill task (not just those with tentacle in the name) as seen in retail screenshots.

Corrects spells, armor levels, and skills of the tentacles to match the pcap, wiki, and screenshots.

Removed weight of the world from eyestalks (this was limited to eyestalks on moar city, and the tthuun boss fight).

Added missing wharu fetish effect to the tthuun altar (fetish spawns above the altar) as seen in retail screenshots.

Revised the final scene (where olthoi, tentacles, and sclavus fight each other) to match retail screenshot.

Corrected appearance of the dungeon portal to match teaser screenshot.