update bloodstone factory (#1773)
* update bloodstone factory

Resolved issue with very slow respawn at the end. The respawn is now 5 min.

Sir Donovan no longer gets stuck in a reading state.

Added about 20 seconds delay between the stockpile and the bloodstone waves, as seen in retail video by Berek.

Adjusted scale of the stockpile based on the video.

* adjust respawn to match wiki

As noted on the wiki, the respawn of the Bloodstone is 15-20 min, so it is now set 15 min (previously, was 20).