Update Rynthid Quests
The Tormented Watcher in Crimson Scarab Corruption quest now spawns the portal on death directly instead of using a local signal. This resolves an issue where the he could be killed out of range of the signal.

Replaces local signal with activate in Seed of Power quest, where killing the Hollow Servitor destroys a door.